Modified preferences and refactored public DialogFragments into own package
