+++ Trail Tracker/src/main/AndroidManifest.xml
@@ -53,6 +53,15 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="net.taptools.android.trailtracker.MainActivity" />
         </activity>
+        <activity
+            android:name=".SettingsActivity"
+            android:label="@string/settings"
+            android:parentActivityName=".MainActivity">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value="net.taptools.android.trailtracker.MainActivity" />
+        </activity>
+
     </application>
 
 </manifest>
