+++ Trail Tracker/src/main/java/net/taptools/android/trailtracker/SettingsActivity.java
@@ -12,8 +12,6 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.os.Build;
 
-
-
 public class SettingsActivity extends ActionBarActivity {
 
     @Override
@@ -25,26 +23,6 @@ public class SettingsActivity extends ActionBarActivity {
                     .add(R.id.container, new SettingsFragment())
                     .commit();
         }
-    }
-
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.settings, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-        if (id == R.id.action_settings) {
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
     }
 
     /**
